+++ src/m8flow_bpmn_core/models/process_model_bpmn_version.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, String, Text, UniqueConstraint
+from sqlalchemy import BIGINT, DateTime, String, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from m8flow_bpmn_core.models.base import Base
@@ -30,7 +30,7 @@ class ProcessModelBpmnVersionModel(M8fTenantScopedMixin, TenantScoped, Base):
     bpmn_xml_hash: Mapped[str] = mapped_column(String(64), nullable=False, index=True)
     bpmn_xml_file_contents: Mapped[str] = mapped_column(Text, nullable=False)
     created_at_in_seconds: Mapped[int] = mapped_column(
-        Integer, nullable=False, index=True
+        BIGINT, nullable=False, index=True
     )
     created_at: Mapped[datetime | None] = mapped_column(
         DateTime(timezone=True), nullable=True
