+++ src/m8flow_bpmn_core/models/task_definition.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import JSON, DateTime, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import BIGINT, JSON, DateTime, ForeignKey, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from m8flow_bpmn_core.models.base import Base
@@ -37,8 +37,8 @@ class TaskDefinitionModel(M8fTenantScopedMixin, TenantScoped, Base):
     bpmn_name: Mapped[str | None] = mapped_column(String(255), index=True)
     typename: Mapped[str] = mapped_column(String(255), index=True, nullable=False)
     properties_json: Mapped[dict[str, Any]] = mapped_column(JSON, nullable=False)
-    updated_at_in_seconds: Mapped[int | None] = mapped_column(Integer)
-    created_at_in_seconds: Mapped[int | None] = mapped_column(Integer)
+    updated_at_in_seconds: Mapped[int | None] = mapped_column(BIGINT)
+    created_at_in_seconds: Mapped[int | None] = mapped_column(BIGINT)
     updated_at: Mapped[datetime | None] = mapped_column(
         DateTime(timezone=True), nullable=True
     )
