+++ src/m8flow_bpmn_core/models/tenant.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from enum import StrEnum
 
-from sqlalchemy import DateTime, Enum, Integer, String
+from sqlalchemy import BIGINT, DateTime, Enum, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from m8flow_bpmn_core.models.base import Base
@@ -43,12 +43,12 @@ class M8flowTenantModel(Base):
         nullable=False,
     )
     created_at_in_seconds: Mapped[int] = mapped_column(
-        Integer,
+        BIGINT,
         default=0,
         nullable=False,
     )
     updated_at_in_seconds: Mapped[int] = mapped_column(
-        Integer,
+        BIGINT,
         default=0,
         nullable=False,
     )
